Extract HasAttributes trait from API response.
